requirements.txt
Flask-SQLAlchemy==3.0.0